src/requirements.txt
ansi2html==1.8.0
APScheduler==3.10.4
blinker==1.7.0
click==8.1.7
Flask==3.0.0
gitdb==4.0.11
GitPython==3.1.40
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
prometheus-client==0.19.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
smmap==5.0.1
tzlocal==5.2
Werkzeug==3.0.1
